requirements.txt

# Additional requirements go here

django-plans
